pyproject.toml
[project]
name = "trading-stock-agents"
version = "0.1.0"
description = ""
authors = [
  {name = "MotiTheWizerd", email = "motielmakayes@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
  "yfinance>=0.2.64,<0.3.0",
  "matplotlib>=3.10.3,<4.0.0",
  "mplfinance>=0.12.10b0,<0.13.0",
  "ta>=0.11.0,<0.12.0",
  "schedule>=1.2.2,<2.0.0",
  "joblib>=1.5.1,<2.0.0",
  "scikit-learn>=1.7.0,<2.0.0",
  "pandas>=2.3.0,<3.0.0",
  "numpy>=2.3.1,<3.0.0",
  "pytz>=2025.2,<2026.0",
  "tqdm>=4.67.1,<5.0.0",
  "pytest>=8.4.1,<9.0.0",
  "seaborn>=0.13.2,<0.14.0",
  "pytest-cov>=6.2.1,<7.0.0",
  "mem0ai>=0.1.114,<0.2.0",
  "google-genai (>=1.24.0,<2.0.0)",
  "qdrant-client (>=1.14.3,<2.0.0)",
  "rich (>=14.0.0,<15.0.0)"
]

[tool.poetry]
packages = [
    { include = "trading", from = "src" },
    { include = "tests", format = "sdist" }
]

[tool.poetry.scripts]
trading = "trading.cli:main"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
pythonpath = [".", "src"]
addopts = "-v --import-mode=importlib"
filterwarnings = [
    "ignore::UserWarning",
    "ignore::pytest.PytestCollectionWarning",
    "ignore::pytest.PytestConfigWarning",
    "ignore::DeprecationWarning",
    "ignore::FutureWarning",
    "ignore::ResourceWarning",
    "ignore:TA-Lib not installed"
]
